Фикс редактирования новостей, добавлено имя модератора, перенесено добавление пользователей в fill_test_users
