+++ BestJob/templates/moderator_profile.html
@@ -6,7 +6,7 @@
 {% endblock %}
 
 {% block header %}
-    {% include 'inc_news_create_header.html' %}
+    {% include 'inc_moderator_header.html' %}
 {% endblock %}
 
 {% block content %}
@@ -24,7 +24,10 @@
                                  <img width="400" height="400"
                                       src="{% if object.image %} {{ object.image.url }} {% else %} {% static 'img/users/default_avatar.jpg' %} {% endif %}" class="img-thumbnail">
                         </div>
-
+                        <div class="form-group">
+                            <label>Имя модератора</label>
+                            {{ form.name }}
+                        </div>
                         <div class="form-group">
                             {{ form.image }}
                             <label class="custom-file-label" for="image">Выберите изображение</label>
