+++ BestJob/templates/news_detail.html
@@ -23,7 +23,7 @@
                     <h2>{{ news_object.title }}</h2>
                     <!-- Post Details -->
                     <div class="post-detail">
-                        <span><i class="fa fa-user"></i>{{ news_object.author.username }}</span>
+                        <span><i class="fa fa-user"></i>{{ news_object.author.name }}</span>
                         <span><i class="fa fa-clock-o"></i>{{ news_object.created|date:"H:i" }}</span>
                     </div>
                 </div>
