+++ BestJob/users/fixtures/moderator.json
@@ -1,5 +1,6 @@
 [
   {
+    "name": "Петр Петров",
     "image": "moderator_photo/1.jpg",
     "date_create": "2022-06-15T14:04:22"
   }
